requirements.txt
anyio==4.0.0
azure-functions==1.17.0
certifi==2023.7.22
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tzdata==2023.3